pyproject.toml
[build-system]
requires = [ "setuptools>=42", "wheel" ]
build-backend = "setuptools.build_meta"

[project]
name = "allconf"
dynamic = ["version"]
description = "Configuration file reader with some nifty bells and whistles added"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [
    { name = "Thordur Matthiasson", email = "doddi79@gmail.com" }
]
keywords = [ "config", "configuration", "json", "yaml", "tools", "utils" ]
classifiers = [
    "Development Status :: 5 - Production/Stable",

    "License :: OSI Approved :: MIT License",

    "Intended Audience :: Developers",

    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",

    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]

dependencies = [
  "PyYAML >=6.0.1, <7",
  "batutils >=2, <3"
]

[project.optional-dependencies]
fidelius = [
    "fidelius >=1.0, <3"
]

[project.scripts]
allconf-render = "allconf.cli.render.main:main"
allconf-stubber = "allconf.cli.stubber.main:main"

[project.urls]
Homepage = "https://github.com/doddi79/allconf"
Documentation = "https://github.com/doddi79/allconf/blob/main/README.md"
Repository = "https://github.com/doddi79/allconf.git"
Issues = "https://github.com/doddi79/allconf/issues"
Changelog = "https://github.com/doddi79/allconf/blob/main/CHANGELOG.md"

[tool.setuptools.dynamic]
version = {attr = "allconf.__version__"}

[tool.setuptools.packages.find]
where = [ "." ]
exclude = [ "tests", "tests.*" ]